modified FSM for simultaneous control
The FSM was not set up to have both motors initialized at the same time. I think we should have the FSM belong to the swivel motor (M2), while the drive motor (M1) just responds to PWM packets. Need to make sure it still stops when the E-stop packet is sent, and that it still times out.